fix: injecting prisma dependency (#39)
* injecting prisma dependency

We had several "too many database connections" errors in production environment. The reason is probably because we created a new Prisma Client everytime we used the database, but we never closed the connection.

Now, I am delegating the lifetime management of the Prisma client to the DI container. We will use the Prisma client per scope, and dispose it by the end of the request